Fix parsing DSN so query string is handled correctly
Previously the query string when parsing the data source was ignored.
This prevented optional settings to be provided such as `useTLS`, which
then did not allow users to sign in to (first time) MySQL when using
`caching_sha2_password`.

We now correctly detect whether there is a query string. This makes now
sure that for example `useTLS` is respected.

We also added missing tests.

Fixes #2.