fix: recover from server timing out connections
When `mysqlx_wait_timeout` is reached, the MySQL X Plugin will close
the inactive session. We did not correctly recover from this. We fix
it by sending the `driver.ErrBadConn` when the header of a message
could not be read.

We also implement `xmysql.UnmarshalPartial` which is the same as
`proto.Unmarshall` but with `AllowPartial` set `true`.

Fixes #22