fix: handling too large packets
The MySQL X Plugin does not report whether a packet is too big. It
instead simply closes the connection and the client cannot figure out
what the problem was.

We fix this by getting the `mysqlx_max_allowed_packet` value and check
the message size before writing it to the server.

Fixes #28