feat(driver!): register drivers pxmysql and mysql
We register both the `pxmysql` and `mysql` driver names. This makes it a
bit easier to use. The driver name is really the "dialect".

If for any reason one wants to use another driver registering 'mysql',
the registration can be disabled using following environment variable
(value must be "1"):

    PXMYSQL_DONT_REGISTER_MYSQL=1

BREAKING CHANGE:
Previously, we used the driver name `mysqlpx`. This was different from
the package name, and since it even confused the author, we renamed
it back as `pxmysql`. Since we have not released a production version
we allow ourselves to be not backwards compatible. Users must adapt.

Fixes #34